+++ src/app/(sub)/join/components/PopupHeader.tsx
