+++ src/app/(sub)/join/components/StepTitle.tsx
