+++ src/app/(sub)/join/components/WaitingHeader.tsx
