+++ src/app/(sub)/join/funnels/step2/Step2.tsx
