+++ src/app/(sub)/join/funnels/step2/components/MbtiBottomSheet.tsx
