+++ src/app/(sub)/join/funnels/step2/components/MoodBoard.tsx
