+++ src/app/(sub)/join/funnels/step2/components/SectionLabel.tsx
