+++ src/app/(sub)/join/funnels/step6/Step6.tsx
