--- src/app/(sub)/moodboard/page.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-export default function MoodBoard() {
-  return (
-    <div>
-      <h1>MoodBoard</h1>
-    </div>
-  );
-}
